fix: corrected some errors
