+++ .env.example
@@ -1,2 +1,3 @@
 ACCESS_TOKEN=8234d078-e3ab-479e-a20e-89eb4dd0133f
 CLIENT=3W6izon01E77goCGve8pHA
+NODE_ENV=development
