+++ src/middleware/parseBody.js
@@ -34,3 +34,4 @@ export default (req, res, next) => {
     }
     next(); // Proceed to the next middleware
 };
+
