+++ src/utils/errorCodesGenerator.js
@@ -1,13 +1,3 @@
-// errorMessages.js (a separate file to store messages)
-export const errorMessages = {
-	isCard: 'creditCardNumber: The card number is invalid',
-	expirationDate: 'expirationDate: The card has expired',
-	validEmail: 'email: The email is not valid',
-	validCvv: 'cVV2: The CVV2 is not valid',
-	validPhoneNumber: 'phoneNumber: The phone number is not valid, only Nigerian phone numbers are allowed',
-};
-
-// errorCodesGenerator.js
 import { errorMessages } from './errorMessages';
 
 export default function generateErrorCodes({
