+++ src/utils/errorMessages.js
@@ -1,0 +1,7 @@
+export const errorMessages = {
+	isCard: 'creditCardNumber: The card number is invalid',
+	expirationDate: 'expirationDate: The card has expired',
+	validEmail: 'email: The email is not valid',
+	validCvv: 'cVV2: The CVV2 is not valid',
+	validPhoneNumber: 'phoneNumber: The phone number is not valid, only Nigerian phone numbers are allowed',
+};
